% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyOfdailytherapyheatmap.R
\name{olddailytherapyheatmap}
\alias{olddailytherapyheatmap}
\title{summarisedata}
\usage{
olddailytherapyheatmap()
}
\arguments{
\item{teamdataframe}{A source tibble containing the aggregated SSNAP data by month for the team. You should 
use the functions in SSNAPInterface to obtain the raw data, and the functions in 
SSNAPStats to aggregate the data appropriately.}

\item{nationaldomaindataframe}{A source tibble containing the national domain results for all teams.}

\item{reportingPeriodStart}{the date the reporting period will start from}

\item{reportingPeriodEnd}{the date the reporting period will end}

\item{pathname}{The pathname for the resulting presentation file(s). The file will be called
Team_XXX_MonthlySlideshow_Jan00-Mar00 where XXX is the TeamCode, and Jan00 and Mar00 are the start and end
date of the current reporting period respectively.}
}
\value{
The presentation is saved as the proposed filename
}
\description{
\code{summarisedata} creates a slideshow for an acute team.
}
\details{
This function takes an imported data table from SSNAP and creates a slideshow of performance
over the previous year to inform teams of their performance.
}
\author{
Andrew Hill, \email{andrew.hill@doctors.org.uk}
}
